Backend/src/aimodel/pyproject.toml
[project]
name = "aimodel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.117.1",
    "librosa>=0.11.0",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pydantic>=2.11.9",
    "python-multipart>=0.0.20",
    "uvicorn>=0.37.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
    "spacy>=3.7.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.1.1",
]
